- Test database schema: 	- Converted decimal to money data types where it made sense (Price, ListPrice, etc) 	- Added image field to Product schema to test byte arrays 	- Added Height, Width, Depth, Weight, and ShippingWeight fields to Product schema typed as decimal (ensures decimal fields as others converted to money) - Better handling of parameter building by using the declared type instead of generic type (T was sometimes object, where declared type provides more correct/concrete type for inference) and accounting for null and DBNull when finding existing parameters - Support for money data type, with clr type of double - Support for varbinary data type, with clr type of byte array - Added concrete expression types for "In" and "Like" expressions and removed corresponding enum values from FilterExpressionOperator - Reworked formatting of where clauses in append process - Added stack concept for FieldExpression in AssemblyContext similar to append styles.  Enusures field hints can be assigned and retained if another field hint is applied for a recursive append process. - Fixed part appender factory by adding generic versions of IAssemblyPartAppender interface(s) to FilterExpressionPartAppender (ensures "as T" when creating an appender of the more concrete version of FilterExpression - i.e. FilterExpression<bool>) results in the appender, not null

+++ test/HatTrick.DbEx.MsSql.Test.Database/Executor/Arithmetic.cs
@@ -50,7 +50,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
 
         [Theory]
         [MsSqlVersions.AllVersions]
-        public void Does_arithmetic_with_no_order_of_precedence_succeed(int version, decimal expected = 1002.00m)
+        public void Does_arithmetic_with_no_order_of_precedence_succeed(int version, double expected = 1002.00)
         {
             //given
             ConfigureForMsSqlVersion(version);
@@ -58,10 +58,10 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
             var exp = db.SelectOne(
                     dbo.Product.ListPrice * dbo.Product.Quantity + 3
                 ).From(dbo.Product)
-                .Where(dbo.Product.ListPrice == 9.99m);
+                .Where(dbo.Product.ListPrice == 9.99);
 
             //when               
-            decimal result = exp.Execute();
+            double result = exp.Execute();
 
             //then
             result.Should().Be(expected);
